+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 # Changelog
 All notable changes to core_tools will be documented in this file.
 
-## \[1.4.7] - 2022-11-@@@
+## \[1.4.7] - 2022-11-18
 
 - Fixed import of NumpyJSONEncoder for new qcodes versions
 - Improved error handling during measurement
 - Fixed VideoMode and DataBrowser 2D plots for pyqtgraph 0.13+
 - Lowered VideoMode refresh rate to better support remote connections
+- Added option silent to measurement and scan_generic
+- Keyboard interrupt in measurement now aborts running script (So you don't have to hit Ctrl-C 50 times..)
 
 ## \[1.4.6] - 2022-10-17
 - Improved logfile format